"vortlisto" parser to extract section, english, esperanto, basic word, transitivity
The app fetches the page if it isn't present.

I'd hoped to build a parser to work from the original since it *looked* pretty regular but the old software that produced the page actually had numerous minor errors that turned out to be fiddly to parse so I opted to hand-fix them (GCSELIST_FIXED_PATH) and build a parser that assumed accurate input.

Output is currently a single CSV.